Add CSS file and set body margin to zero
This addresses an issue with a scroll bar that was appearing in the app. The hook can only look at the App element's height but not the full body, which had some margin in this case. We should fix the hook in the future to allow for this use case.